always set non-null writer batch size
